feat(egh): enhance post removal with egghead playlist sync
