+++ apps/egghead/src/app/(content)/posts/_components/resource-resources-list.tsx
@@ -16,13 +16,12 @@ import {
 	createPost,
 	getPost,
 	removeEggheadLessonFromPlaylist,
+	removePostFromCoursePost,
 } from '@/lib/posts-query'
 import { createResource } from '@/lib/resources/create-resources'
 
 import type { ContentResource } from '@coursebuilder/core/schemas'
 import { Button } from '@coursebuilder/ui'
-
-import { removePostFromCoursePost } from '../actions'
 
 type FormState = {
 	activeForm: 'lesson' | 'section' | 'existing_lesson' | null
@@ -130,21 +129,10 @@ export function ResourceResourcesList({
 				rootResourceId={resource.id}
 				state={state}
 				updateState={updateState}
-				onDelete={async ({ itemId }: { itemId: string }) => {
-					const post = await getPost(itemId)
-
-					if (!post || !post.fields?.eggheadLessonId) {
-						throw new Error('eggheadLessonId is required')
-					}
-
+				onItemDelete={async ({ itemId }: { itemId: string }) => {
 					await removePostFromCoursePost({
 						postId: itemId,
 						resourceOfId: resource.id,
-					})
-
-					await removeEggheadLessonFromPlaylist({
-						eggheadLessonId: post.fields?.eggheadLessonId,
-						eggheadPlaylistId: resource.fields?.eggheadPlaylistId,
 					})
 				}}
 			/>
