--- apps/egghead/src/app/(content)/posts/actions.ts
@@ -1,22 +1,0 @@
-'use server'
-
-import { db } from '@/db'
-import { contentResourceResource } from '@/db/schema'
-import { and, eq } from 'drizzle-orm'
-
-export async function removePostFromCoursePost({
-	postId,
-	resourceOfId,
-}: {
-	postId: string
-	resourceOfId: string
-}) {
-	return await db
-		.delete(contentResourceResource)
-		.where(
-			and(
-				eq(contentResourceResource.resourceOfId, resourceOfId),
-				eq(contentResourceResource.resourceId, postId),
-			),
-		)
-}
